move findThought into real code
